+++ src/engine/info.h
@@ -962,8 +962,8 @@ typedef enum {
 
 typedef struct {
 	spritenum_t     sprite;
-	long            frame;
-	long            tics;
+	intptr_t            frame;
+	intptr_t            tics;
 	actionf_t       action;
 	statenum_t      nextstate;
 	int		misc1, misc2;
