+++ src/engine/m_misc.c
@@ -245,9 +245,9 @@ int M_ReadFile(int8_t const* name, byte** buffer) {
 // M_FileLength
 //
 
-long M_FileLength(FILE* handle) {
-	long savedpos;
-	long length;
+intptr_t M_FileLength(FILE* handle) {
+	intptr_t savedpos;
+	intptr_t length;
 
 	// save the current position in the file
 	savedpos = ftell(handle);
